Comment revision for the previous

+++ speeduino/corrections.ino
@@ -321,7 +321,7 @@ uint16_t correctionAccel()
               int16_t taperRange = (int16_t) configPage2.aeColdTaperMax - configPage2.aeColdTaperMin;
               int16_t taperPercent = (int)((currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET - configPage2.aeColdTaperMin) * 100) / taperRange;
               int16_t coldPct = (int16_t) 100+ percentage((100-taperPercent), configPage2.aeColdPct-100);
-              unsigned long accelValue_long = (unsigned long) accelValue * coldPct / 100; //Potential overflow (if AE is large) without using unsigned long
+              unsigned long accelValue_long = (unsigned long) accelValue * coldPct / 100; //Potential overflow (if AE is large) without using uint16_t
               accelValue = (uint16_t) accelValue_long;
             }
           }
@@ -383,7 +383,7 @@ uint16_t correctionAccel()
               int16_t taperRange = (int16_t) configPage2.aeColdTaperMax - configPage2.aeColdTaperMin;
               int16_t taperPercent = (int)((currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET - configPage2.aeColdTaperMin) * 100) / taperRange;
               int16_t coldPct = (int16_t) 100+ percentage((100-taperPercent), configPage2.aeColdPct-100);
-              uint16_t accelValue_uint = (uint16_t) accelValue * coldPct / 100; //Potential overflow (if AE is large) without using unsigned long
+              uint16_t accelValue_uint = (uint16_t) accelValue * coldPct / 100; //Potential overflow (if AE is large) without using uint16_t
               accelValue = (uint16_t) accelValue_uint;
             }
           }
